+++ src/main/java/com/blackrook/rookscript/resolvers/variable/AbstractVariableResolver.java
@@ -223,7 +223,7 @@ public class AbstractVariableResolver implements ScriptVariableResolver
 		@Override
 		public String toString()
 		{
-			return name + ": " + value;
+			return "\"" + name + "\": " + value;
 		}
 		
 	}